[pkg]-use float time for movement

+++ src/BuiltInPackages/LinqSTG/LuaSTGEditorSharpV2.Package.LinqSTG.Window/PointPrediction.cs
@@ -4,7 +4,7 @@ using System.Numerics;
 namespace LuaSTGEditorSharpV2.Package.LinqSTG.Windows
 {
     public readonly record struct PointPrediction(
-        IParametric<int, Vector2> PointFunc,
+        IParametric<float, Vector2> PointFunc,
         int StartTime,
         BulletShape Shape,
         float Diameter)
