+++ src/BuiltInPackages/LinqSTG/LuaSTGEditorSharpV2.Package.LinqSTG.Window/ViewModel/Nodes/MovementOperator/MovementOffsetNode.cs
@@ -9,9 +9,9 @@ namespace LuaSTGEditorSharpV2.Package.LinqSTG.Windows.ViewModel.Nodes.MovementOp
 {
     public class MovementOffsetNode : LinqSTGNodeViewModel
     {
-        public LinqSTGNodeInputViewModel<Contextual<IParametric<int, Vector2>>?> InputMovement { get; }
+        public LinqSTGNodeInputViewModel<Contextual<IParametric<float, Vector2>>?> InputMovement { get; }
         public LinqSTGNodeInputViewModel<Contextual<Vector2>?> InputOffset { get; }
-        public LinqSTGNodeOutputViewModel<Contextual<IParametric<int, Vector2>>> OutputMovement { get; }
+        public LinqSTGNodeOutputViewModel<Contextual<IParametric<float, Vector2>>> OutputMovement { get; }
 
         public MovementOffsetNode()
         {
@@ -30,7 +30,7 @@ namespace LuaSTGEditorSharpV2.Package.LinqSTG.Windows.ViewModel.Nodes.MovementOp
             OutputMovement.Value = InputMovement.ValueChanged
                 .CombineLatest(InputOffset.ValueChanged, (movement, offset)
                     => Contextual.Create(dict
-                        => new Parametric<int, Vector2>(t
+                        => new Parametric<float, Vector2>(t
                             => (movement?.Invoke(dict)?.Predict(t) ?? Vector2.Zero)
                                 + (offset?.Invoke(dict) ?? Vector2.Zero))));
         }
